requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Core dependencies
setuptools>=65.0.0
wheel>=0.37.0

# Numerical computing
numpy>=1.21.0,<2.0.0

# ONNX Runtime (required for InsightFace)
onnxruntime>=1.15.0

# Face recognition
insightface==0.7.3
opencv-python-headless>=4.8.0
scikit-learn>=1.3.0

# Image processing
Pillow>=10.0.0

# Data validation
pydantic==2.5.2
pydantic-settings==2.1.0

# File handling
aiofiles==23.2.0
python-multipart==0.0.6

# Templates
jinja2==3.1.2

# Logging
structlog==23.2.0

# Utils
python-dotenv==1.0.0